requirements.txt
pillow
watchfiles